Refactor board related components into separate files
